(feat) Add notification counter
