+++ packages/esm-patient-notifications-app/src/notification-action-button/notification-action-button.extension.tsx
@@ -2,12 +2,19 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import { ActionMenuButton } from "@openmrs/esm-framework";
 import { Notification } from "@carbon/react/icons";
-import { useLaunchWorkspaceRequiringVisit } from "@openmrs/esm-patient-common-lib";
+import {
+  useLaunchWorkspaceRequiringVisit,
+  getPatientUuidFromStore,
+} from "@openmrs/esm-patient-common-lib";
+import usePatientNotifications from "../hooks/usePatientNotifications";
 
 const NotificationActionButton: React.FC = () => {
   const { t } = useTranslation();
   const launchNotificationWorkSpace =
     useLaunchWorkspaceRequiringVisit("notifications");
+
+  const patientUuid = getPatientUuidFromStore();
+  const { notifications } = usePatientNotifications(patientUuid);
 
   return (
     <ActionMenuButton
@@ -16,6 +23,7 @@ const NotificationActionButton: React.FC = () => {
       iconDescription={t("notifications", "Notifications")}
       handler={launchNotificationWorkSpace}
       type={"button"}
+      tagContent={notifications?.length > 0 ? notifications?.length : null}
     />
   );
 };
