+++ packages/esm-patient-notifications-app/src/notification-component/notification.scss
@@ -101,3 +101,8 @@
 .mustardTag{
   color: colors.$yellow-50;
 }
+
+.loading {
+  padding: 3rem;
+  font-weight: 600;
+}
