+++ packages/esm-patient-notifications-app/src/patient-flags-components/patient-flag.scss
@@ -33,6 +33,7 @@
 }
 
 .loading {
+  margin-left: 10rem;
   padding: 3rem;
   font-weight: 600;
 }
